requirements.txt
setuptools==67.7.2
pandas==2.0.1
polars==0.17.11
pyarrow~=11.0.0
pysam~=0.20.0
biopython~=1.79
numpy~=1.24.2
matplotlib~=3.7.1
seaborn~=0.12.2
plotly~=5.14.1
interval~=1.0.0
matplotlib-venn~=0.11.9
seqlogo~=5.29.8
scipy~=1.10.1
statsmodels~=0.13.2
scikit-learn~=1.2.2
rpy2~=3.5.1
mygene~=3.2.2
pyyaml~=6.0
requests~=2.28.2
xlsxwriter~=3.1.0
fastcluster~=1.2.6
kaleido~=0.2.1
